Using strongly typed approach.

+++ ConsoleApplication1/Program.cs
@@ -9,7 +9,7 @@ namespace ConsoleApplication1
         {
             BCM2835.bcm2835_init();
             {
-                BCM2835.bcm2835_gpio_fsel((byte)BCM2835.RPiGPIOPin.RPI_V2_GPIO_P1_11, (byte)BCM2835.bcm2835FunctionSelect.BCM2835_GPIO_FSEL_OUTP);
+                BCM2835.bcm2835_gpio_fsel(BCM2835.RPiGPIOPin.RPI_V2_GPIO_P1_11, BCM2835.bcm2835FunctionSelect.BCM2835_GPIO_FSEL_OUTP);
                 Console.WriteLine("Press <L> set to set pin to LOW or <H> to HIGH and <Q> to quit.");
                 while (true)
                 {
@@ -17,9 +17,9 @@ namespace ConsoleApplication1
                     if (key == ConsoleKey.Q)
                         break;
                     else if (key == ConsoleKey.L)
-                        BCM2835.bcm2835_gpio_clr((byte)BCM2835.RPiGPIOPin.RPI_V2_GPIO_P1_11);
+                        BCM2835.bcm2835_gpio_clr(BCM2835.RPiGPIOPin.RPI_V2_GPIO_P1_11);
                     else if (key == ConsoleKey.H)
-                        BCM2835.bcm2835_gpio_set((byte)BCM2835.RPiGPIOPin.RPI_V2_GPIO_P1_11);
+                        BCM2835.bcm2835_gpio_set(BCM2835.RPiGPIOPin.RPI_V2_GPIO_P1_11);
 
                 }
             }
